doccoon_backend/pyproject.toml
[project]
name = "doccoon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bleach>=6.3.0",
    "cryptography>=46.0.4",
    "django>=6.0",
    "django-allauth>=65.14.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.1.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "google-genai>=1.61.0",
    "gunicorn>=23.0.0",
    "openai>=2.16.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]
